+++ packages/ts-newskit/src/utils/index.tsx
@@ -6,3 +6,6 @@ export { convertDateToWeekday } from './date-formatting';
 export { formatThemeOverrides } from './theme-formatting';
 export { ContentWrapper } from './contentWrapper';
 export { renderComponent } from './responsiveRender';
+export { getResponsiveNavData } from './getResponsiveNavData';
+export { getActiveArticleFlags } from './getActiveArticleFlag';
+export { getForcedExternalContentRatio } from './getForcedExternalContentRatio';