+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.49.53](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.52...@times-components/ssr@2.49.53) (2024-01-16)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.49.52](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.51...@times-components/ssr@2.49.52) (2024-01-16)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.49.51](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.50...@times-components/ssr@2.49.51) (2024-01-15)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.49.50](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.49...@times-components/ssr@2.49.50) (2024-01-15)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.49.49](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.48...@times-components/ssr@2.49.49) (2024-01-12)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
+## [2.49.48](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.47...@times-components/ssr@2.49.48) (2024-01-12)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.49.47](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.49.46...@times-components/ssr@2.49.47) (2024-01-12)
 
 **Note:** Version bump only for package @times-components/ssr